rag-service/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
numpy
pytest
google-generativeai
beautifulsoup4